# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Nour — Egyptian Fashion & Beauty Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryBody"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    put:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCategoryBody"
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "404":
          description: Not found
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        nameAr:
          type: string
          description: Arabic name
        parentId:
          type: integer
          nullable: true
          description: ID of the parent category
        type:
          type: string
          enum: [fashion, cosmetics]
        imageUrl:
          type: string
          nullable: true
        productCount:
          type: integer
      required: [id, name, nameAr, type, productCount]
    CreateCategoryBody:
      type: object
      properties:
        name:
          type: string
        nameAr:
          type: string
        parentId:
          type: integer
          nullable: true
        type:
          type: string
          enum: [fashion, cosmetics]
        imageUrl:
          type: string
          nullable: true
      required: [name, nameAr, type]
    UpdateCategoryBody:
      type: object
      properties:
        name:
          type: string
        nameAr:
          type: string
        parentId:
          type: integer
          nullable: true
        type:
          type: string
          enum: [fashion, cosmetics]
        imageUrl:
          type: string
          nullable: true
...